Change EventNumber.DeletedStream to int.MaxValue rather than long.MaxValue.

+++ src/EventStore.Core/Data/EventNumber.cs
@@ -2,7 +2,7 @@ namespace EventStore.Core.Data
 {
     public static class EventNumber
     {
-        public const long DeletedStream = long.MaxValue;
+        public const long DeletedStream = int.MaxValue;
         public const long Invalid = int.MinValue;
     }
 }
